+++ spinta/datasets/backends/dataframe/ufuncs/query/ufuncs.py
@@ -74,6 +74,15 @@ def offset(env: DaskDataFrameQueryBuilder, n: int):
 
 @ufunc.resolver(DaskDataFrameQueryBuilder, GetAttr)
 def _resolve_property(env: DaskDataFrameQueryBuilder, attr: GetAttr) -> Property:
+    obj = str(attr.obj)
+    name = str(attr.name)
+    if obj in env.model.properties:
+        prop = env.model.properties.get(obj)
+        dtype = getattr(prop, "dtype", None)
+        langs = getattr(dtype, "langs", None)
+        if isinstance(dtype, Text) and langs:
+            if name in langs:
+                return env.call("_resolve_property", langs[name])
     return env.call("_resolve_property", attr.obj)
 
 
@@ -127,8 +136,30 @@ def select(env: DaskDataFrameQueryBuilder, expr: Expr):
 
     env.selected = {}
     if args:
+        resolved = {}
+        selected_keys = set()
+        selected_languages = {}
         for key, arg in args:
-            env.selected[key] = env.call("select", arg)
+            resolved[key] = env.call("_resolve_property", arg)
+        for key, prop in resolved.items():
+            prop_parent = getattr(prop, "parent", None)
+            if prop_parent and isinstance(prop_parent.dtype, Text):
+                selected_keys.add(prop_parent.place)
+                selected_languages.setdefault(prop_parent.place, set()).add(prop.name)
+            else:
+                selected_keys.add(prop.name)
+        for selected_key in selected_keys:
+            prop = env.model.flatprops.get(selected_key)
+            if isinstance(prop.dtype, Text) and authorized(env.context, prop, Action.GETALL):
+                selected_language = selected_languages.get(prop.place)
+                if not selected_language:
+                    env.selected[prop.place] = env.call("select", prop)
+                else:
+                    env.selected[prop.place] = env.call("select", prop, selected_language)
+            elif authorized(env.context, prop, Action.GETALL):
+                env.selected[resolved[selected_key].place] = env.call("select", resolved[selected_key])
+            else:
+                raise PropertyNotFound(env.model, property=resolved[selected_key])
     else:
         for prop in take(["_id", all], env.model.properties).values():
             if authorized(env.context, prop, Action.GETALL):
@@ -166,6 +197,11 @@ def _get_property_for_select(
         return prop
     else:
         raise PropertyNotFound(env.model, property=name)
+
+
+@ufunc.resolver(DaskDataFrameQueryBuilder, Property, set)
+def select(env: DaskDataFrameQueryBuilder, prop: Property, keys: set) -> Selected:
+    return env.call("select", prop.dtype, keys)
 
 
 @ufunc.resolver(DaskDataFrameQueryBuilder, Property)
@@ -209,6 +245,17 @@ def select(env: DaskDataFrameQueryBuilder, prop: Property) -> Selected:
         assert isinstance(result, Selected), prop
         env.resolved[prop.place] = result
     return env.resolved[prop.place]
+
+
+@ufunc.resolver(DaskDataFrameQueryBuilder, Text, set)
+def select(env: DaskDataFrameQueryBuilder, dtype: Text, languages: set) -> Selected:
+    prep = {}
+    for lang in languages:
+        if lang in dtype.langs:
+            prep[lang] = env.call("select", dtype.langs[lang])
+        else:
+            raise PropertyNotFound(dtype.prop.model, property=dtype.prop, lang=lang)
+    return Selected(prop=dtype.prop, prep=prep)
 
 
 @ufunc.resolver(DaskDataFrameQueryBuilder, DataType)